app/requirements.txt
yfinance
pandas
azure-identity
azure-storage-blob
pyodbc
alpaca-py
requests
openpyxl
pytz 